pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "tingbok"
dynamic = ["version"]
description = "Product and category lookup service for domestic inventory systems"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "AGPL-3.0-or-later"}
authors = [{ name = "Tobias Brox", email = "tobias@tobixen.no" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Framework :: FastAPI",
]
dependencies = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
skos = ["niquests>=3.0", "pyoxigraph>=0.3"]
ean = ["niquests>=3.0"]
dev = ["pytest>=8.0", "anyio[trio]>=4.0", "httpx>=0.25", "ruff>=0.4", "pre-commit>=4.0"]

[project.urls]
Repository = "https://github.com/tobiasbp/tingbok"
Issues = "https://github.com/tobiasbp/tingbok/issues"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/tingbok/_version.py"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "PT"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["tingbok"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = ["error"]
